Remove QueryAggregateFunctions.java
