+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/aggregatable/count/AggregateCountColumn.java
@@ -26,7 +26,6 @@ import org.bithon.component.commons.expression.IDataType;
 import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
 import org.bithon.server.storage.datasource.column.aggregatable.IAggregatableColumn;
 import org.bithon.server.storage.datasource.query.ast.QueryAggregateFunction;
-import org.bithon.server.storage.datasource.query.ast.QueryAggregateFunctions;
 
 /**
  * @author frank.chen021@outlook.com
@@ -49,7 +48,7 @@ public class AggregateCountColumn implements IAggregatableColumn {
                                 @JsonProperty("alias") @Nullable String alias) {
         this.name = name;
         this.alias = alias == null ? name : alias;
-        this.queryStageAggregator = new QueryAggregateFunctions.CountAggregateExpression(name);
+        this.queryStageAggregator = new QueryAggregateFunction.CountAggregateExpression(name);
     }
 
     @Override
