+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/aggregatable/max/AggregateMaxColumn.java
@@ -20,7 +20,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import org.bithon.server.storage.datasource.column.aggregatable.IAggregatableColumn;
 import org.bithon.server.storage.datasource.query.ast.QueryAggregateFunction;
-import org.bithon.server.storage.datasource.query.ast.QueryAggregateFunctions;
 
 /**
  * @author frank.chen021@outlook.com
@@ -40,7 +39,7 @@ public abstract class AggregateMaxColumn implements IAggregatableColumn {
                               String alias) {
         this.name = name;
         this.alias = alias == null ? name : alias;
-        this.aggregateExpression = new QueryAggregateFunctions.MaxAggregateExpression(name);
+        this.aggregateExpression = new QueryAggregateFunction.MaxAggregateExpression(name);
     }
 
     @JsonIgnore
