+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/aggregatable/sum/AggregateSumColumn.java
@@ -20,7 +20,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import org.bithon.server.storage.datasource.column.aggregatable.IAggregatableColumn;
 import org.bithon.server.storage.datasource.query.ast.QueryAggregateFunction;
-import org.bithon.server.storage.datasource.query.ast.QueryAggregateFunctions;
 
 /**
  * @author frank.chen021@outlook.com
@@ -41,7 +40,7 @@ public abstract class AggregateSumColumn implements IAggregatableColumn {
                               String alias) {
         this.name = name;
         this.alias = alias == null ? name : alias;
-        this.aggregateExpression = new QueryAggregateFunctions.SumAggregateExpression(name);
+        this.aggregateExpression = new QueryAggregateFunction.SumAggregateExpression(name);
     }
 
     @JsonIgnore
